remove SingleNodeSyncGossip
